requirements.txt
playwright>=1.50,<2.0
playwright-stealth>=2.0,<3.0
python-dotenv>=1.0
schedule>=1.2
winotify>=1.1
